requirements.txt
# --------- pytorch --------- #
# torch>=1.10.0
# torchvision>=0.11.0
pytorch-lightning==1.7.7
torchmetrics>=0.7.0

# --------- hydra --------- #
hydra-core>=1.1.0
hydra-colorlog>=1.1.0

# --------- loggers --------- #
wandb
# neptune-client
# mlflow
# comet-ml
# tensorboard

# --------- linters --------- #
pre-commit      # hooks for applying linters on commit
black           # code formatting
ruff            # code analysis
nbstripout      # remove output from jupyter notebooks

# --------- others --------- #
rich            # beautiful text formatting in terminal
pytest          # tests
sh              # for running bash commands in some tests
pudb            # debugger
coolname        # names for trained YOLOv4 model

# --------- darknet --------- #
tensorboardX>=2.0
scikit-image==0.19.2
matplotlib==3.5.1
tqdm==4.62.3
easydict==1.9
Pillow==9.0.0
opencv_python==4.5.5.64
pycocotools

# --------- dir ---------- #
pyro-ppl==1.8.0
ipykernel==6.21.3
ipywidgets==8.0.4
scikit-learn==1.2.1
pandas==1.5.3
seaborn==0.12.2
umap-learn==0.5.3
